extracted access-config-context.ts
